Add input validation and replace Double with BigDecimal for prices
- Add spring-boot-starter-validation dependency
- Add validation annotations to EventRequest and TicketRequest DTOs
- Add @Valid to controller methods for request body validation
- Add validation exception handlers to GlobalExceptionHandler
- Replace Double with BigDecimal for all price/amount fields
- Add repomix to DevContainer

+++ src/main/java/com/example/ticketreservation/dto/TicketResponse.java
@@ -1,6 +1,7 @@
 package com.example.ticketreservation.dto;
 
 import com.example.ticketreservation.entity.TicketStatus;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Value;
@@ -17,7 +18,7 @@ public class TicketResponse {
     String customerName;
     String customerEmail;
     Integer numberOfSeats;
-    Double totalAmount;
+    BigDecimal totalAmount;
     TicketStatus status;
     LocalDateTime createdAt;
     LocalDateTime updatedAt;
